requirements.txt
# ============================================================================
# Enterprise AI Assistant - Complete Requirements
# All dependencies needed to run the entire project
# ============================================================================

# ============================================================================
# Core Dependencies
# ============================================================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# ============================================================================
# AI/ML & LLM Integration
# ============================================================================
openai==1.54.4
tiktoken==0.5.2
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5

# Optional: Anthropic Claude (uncomment if needed)
# anthropic==0.18.1

# ============================================================================
# Vector Store & Embeddings
# ============================================================================
faiss-cpu==1.9.0
sentence-transformers==2.3.1
numpy==1.26.2

# Alternative vector stores (optional)
# pinecone-client==2.2.4
# chromadb==0.4.22

# ============================================================================
# Document Processing
# ============================================================================
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
Pillow==10.1.0

# OCR support (optional)
# pytesseract==0.3.10
# pdf2image==1.16.3

# ============================================================================
# NLP Libraries
# ============================================================================
nltk==3.8.1
langdetect==1.0.9

# Optional: spaCy (if using local NLP models)
# spacy==3.7.2

# ============================================================================
# AWS Integration
# ============================================================================
boto3==1.34.10
botocore==1.34.10

# ============================================================================
# API & Web Framework
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0

# AWS Lambda support
mangum==0.17.0

# ============================================================================
# Frontend Dashboard
# ============================================================================
streamlit==1.29.0
plotly==5.18.0

# ============================================================================
# Data Processing & Utilities
# ============================================================================
tqdm==4.66.1
urllib3==2.1.0

# ============================================================================
# Development & Testing (Optional)
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1
jupyter==1.0.0

# ============================================================================
# Database (Optional - for future use)
# ============================================================================
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# redis==5.0.1

# ============================================================================
# Monitoring & Logging (Optional)
# ============================================================================
# sentry-sdk==1.39.1
# prometheus-client==0.19.0

# ============================================================================
# Version Notes
# ============================================================================
# Python Version: 3.10+
# Last Updated: October 2025
# Total Dependencies: ~35 (core) + ~10 (optional dev tools)
#
# Installation:
#   pip install -r requirements.txt
#
# For development:
#   pip install -r requirements.txt -e .
#
# Minimal install (API only, no document processing):
#   pip install fastapi uvicorn openai langchain faiss-cpu
# ============================================================================